addressed issue #222
