+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/implementations/AbstractProducer.java
@@ -38,7 +38,8 @@ import ca.concordia.encs.citydata.core.utils.RequestOptions;
  * @since 2025-05-27
  */
 
-public abstract class AbstractProducer<E> extends AbstractEntity implements IProducer<E> {
+public sealed abstract class AbstractProducer<E> extends AbstractEntity implements IProducer<E> permits JSONProducer,
+		CSVProducer, ExceptionProducer, FirebaseProducer, PortfolioManagerProducer, PortfolioManagerMetadataProducer {
 	private String filePath;
 	private RequestOptions fileOptions;
 	private IOperation<E> operation;
@@ -79,8 +80,8 @@ public abstract class AbstractProducer<E> extends AbstractEntity implements IPro
 		return runners;
 	}
 
-	public void setResult(ArrayList<E> result) {
-		this.result = result;
+	public void setResult(ArrayList<?> result) {
+		this.result = (ArrayList<E>) result;
 	}
 
 	@Override
