+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/implementations/JSONProducer.java
@@ -17,16 +17,16 @@ import ca.concordia.encs.citydata.core.utils.RequestOptions;
  * @since 2024-12-01
  */
 
-public class JSONProducer extends AbstractProducer<JsonObject> implements IProducer<JsonObject> {
+public non-sealed class JSONProducer extends AbstractProducer<JsonObject> implements IProducer<JsonObject> {
 
 	public JSONProducer(final String filePath, final RequestOptions fileOptions) {
 		super(filePath, fileOptions);
 	}
-	
+
 	public JSONProducer(final String filePath) {
 		super(filePath);
 	}
-	
+
 	@Override
 	public void fetch() {
 		final ArrayList<JsonObject> jsonOutput = new ArrayList<>();
